fix: #WB2-1530, fix feedback PR
