+++ frontend/src/models/page.ts
@@ -12,7 +12,3 @@ export interface Page {
   lastContributerName?: string;
   comments?: Comment[];
 }
-
-export interface ArrayPage {
-  pages: Array<Page>;
-}
